General "requests" infrastructure that should let us consolidate all sorts of objects that require requests from certain people.
For instance, candidate RSVPs need to be confirmed by the VP, officer challenges need to be confirmed by the officer, 
exam files need to be confirmed/approved by the exam files officer, resumes by the indrel officers, etc.

This basic infrastructure has support for RSVPs, should be extended to support more so that it is actually useful.